Fix detection of submodule remote versions
A new method get_submodule_versions_from_remote() is added
to the git module and used to detect the remote revisions
of the configured submodules.

The new method loops over the detected submodules and determines the
remote branch of each submodule as configured in the .gitmodules
file. The default remains "master".